requirements/base.txt
# Base
pytz==2021.3
python-slugify==5.0.2
Pillow==9.0.0
psycopg2==2.9.3 --no-binary psycopg2

# Django
django==4.0.1

# Utilities
django-countries==7.2.1


# Environment
django-environ==0.8.1

# Django REST Framework
djangorestframework==3.13.1
django-filter==2.4.0

# JWT
pyjwt==2.3.0

# Passwords security
argon2-cffi==21.3.0

# Static files
whitenoise==5.3.0

# Celery
tornado==4.2
redis==4.1.0
django-redis==5.2.0
celery==4.4.2
flower==0.9.2
